Added compatibility with Linux and Windows in addition to OSX.
Added ability for Linux and Windows to run this script. Now everyone can use this!
Added friendly and instructive error message for when pypng is not installed.
Fixed bugs that prevented the proper printing output.
Fixed the try/except/finally's to be try/except's.
Fixed inconsistent space and tab indentation which will fail with Python 3.
Fixed sys.exit(0) to exit with code of 1, which indicates an error to the shell.
Added the shebang to the first line so that the script can be run on its own.
Fixed the ImageMagick command line to be more compatible running under Windows shell.